project/capstone-city-scrapper/requirements.txt
beautifulsoup4
mongoengine
python-dotenv
selenium
requests